Applied reviewer comments: fixed some doc strings and renamed several files and classes.
