+++ python/lsst/ts/Dome/llc_configuration_limits/common_amcs_and_lwscs_limits.py
@@ -1,11 +1,11 @@
 from abc import abstractmethod
-from .abstract_llc_configuration_limits import AbstractLlcConfigurationLimits
+from .abstract_limits import AbstractLimits
 import math
 
 DEGREES_TO_RADIANS = math.pi / 180.0
 
 
-class CommonAmcsAndLwsCsConfigurationParameters(AbstractLlcConfigurationLimits):
+class CommonAmcsAndLwscsLimits(AbstractLimits):
     @abstractmethod
     def validate(self, configuration_parameters):
         pass
