+++ python/lsst/ts/Dome/llc_configuration_limits/lwscs_limits.py
@@ -1,9 +1,7 @@
-from .common_amcs_and_lwscs_configuration_limits import (
-    CommonAmcsAndLwsCsConfigurationParameters,
-)
+from .common_amcs_and_lwscs_limits import CommonAmcsAndLwscsLimits
 
 
-class LwscsConfigurationLimits(CommonAmcsAndLwsCsConfigurationParameters):
+class LwscsLimits(CommonAmcsAndLwscsLimits):
     """This class holds the limits of the configuration values for the LWSCS lower level component. It will
     validate any set of configuration parameters againt these limits. It will also convert any
     configuration parameter values expressed in deg/s^n to rad/s^n (with n = 1, 2 or 3).
