+++ python/lsst/ts/Dome/mock_llc_statuses/lcs_status.py
