+++ python/lsst/ts/Dome/mock_llc_statuses/moncs_status.py
