+++ python/lsst/ts/Dome/mock_llc_statuses/thcs_status.py
